Refactored Imanee\Image to ResourceProvider\ImagickResource
