+++ src/Model/ResourceProviderInterface.php
@@ -1,0 +1,13 @@
+<?php
+/**
+ * ResourceProvider Interface
+ * Defines the common Image Resource methods for Imanee
+ */
+
+namespace Imanee\Model;
+
+
+interface ResourceProviderInterface
+{
+
+}
